Refactoring files to segment filesystem stuff into fs_util.cpp
