+++ depends/packages/packages.mk
@@ -1,4 +1,4 @@
-packages:=rocksdb boost openssl libevent zeromq
+packages:= boost openssl libevent zeromq
 native_packages := native_ccache
 
 qt_native_packages =
