+++ src/addrdb.h
@@ -3,71 +3,10 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#ifndef BITCOIN_ADDRDB_H
-#define BITCOIN_ADDRDB_H
+#pragma once
 
+#include "ban.h"
 #include "fs.h"
-#include "serialize.h"
-
-#include <map>
-#include <string>
-
-class CSubNet;
-class CAddrMan;
-class CDataStream;
-class CChainParams;
-
-typedef enum BanReason {
-    BanReasonUnknown = 0,
-    BanReasonNodeMisbehaving = 1,
-    BanReasonManuallyAdded = 2
-} BanReason;
-
-class CBanEntry {
-public:
-    static const int CURRENT_VERSION = 1;
-    int nVersion;
-    int64_t nCreateTime;
-    int64_t nBanUntil;
-    uint8_t banReason;
-
-    CBanEntry() { SetNull(); }
-
-    explicit CBanEntry(int64_t nCreateTimeIn) {
-        SetNull();
-        nCreateTime = nCreateTimeIn;
-    }
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream &s, Operation ser_action) {
-        READWRITE(this->nVersion);
-        READWRITE(nCreateTime);
-        READWRITE(nBanUntil);
-        READWRITE(banReason);
-    }
-
-    void SetNull() {
-        nVersion = CBanEntry::CURRENT_VERSION;
-        nCreateTime = 0;
-        nBanUntil = 0;
-        banReason = BanReasonUnknown;
-    }
-
-    std::string banReasonToString() const {
-        switch (banReason) {
-            case BanReasonNodeMisbehaving:
-                return "node misbehaving";
-            case BanReasonManuallyAdded:
-                return "manually added";
-            default:
-                return "unknown";
-        }
-    }
-};
-
-typedef std::map<CSubNet, CBanEntry> banmap_t;
 
 /** Access to the (IP) address database (peers.dat) */
 class CAddrDB {
@@ -93,5 +32,3 @@ public:
     bool Write(const banmap_t &banSet);
     bool Read(banmap_t &banSet);
 };
-
-#endif // BITCOIN_ADDRDB_H
