+++ src/bitcoind.cpp
@@ -11,13 +11,12 @@
 #include "clientversion.h"
 #include "compat.h"
 #include "config.h"
-#include "fs.h"
 #include "httprpc.h"
 #include "httpserver.h"
 #include "init.h"
 #include "noui.h"
 #include "rpc/server.h"
-#include "util.h"
+#include "fs_util.h"
 #include "utilstrencodings.h"
 #include "walletinitinterface.h"
 #include "support/allocators/secure.h"
