+++ src/fs.h
@@ -14,8 +14,8 @@
 #define NO_BOOST_FILESYSTEM
 #else
 #include <boost/filesystem.hpp>
-#include <boost/filesystem/detail/utf8_codecvt_facet.hpp>
-#include <boost/filesystem/fstream.hpp>
+//#include <boost/filesystem/detail/utf8_codecvt_facet.hpp>
+//#include <boost/filesystem/fstream.hpp>
 #include <boost/filesystem/path.hpp>
 #endif
