+++ src/net.h
@@ -7,7 +7,7 @@
 #ifndef BITCOIN_NET_H
 #define BITCOIN_NET_H
 
-#include "addrdb.h"
+#include "ban.h"
 #include "addrman.h"
 #include "amount.h"
 #include "bloom.h"
