+++ src/qt/clientmodel.cpp
@@ -16,7 +16,7 @@
 #include "net.h"
 #include "txmempool.h"
 #include "ui_interface.h"
-#include "util.h"
+#include "fs_util.h"
 #include "validation.h"
 #include "warnings.h"
