+++ src/random.cpp
@@ -4,6 +4,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "random.h"
+#include "fs_util.h"
 
 #include "crypto/sha512.h"
 #include "support/cleanse.h"
