+++ src/rpc/protocol.h
@@ -6,8 +6,6 @@
 #ifndef BITCOIN_RPCPROTOCOL_H
 #define BITCOIN_RPCPROTOCOL_H
 
-#include "fs.h"
-
 #include <cstdint>
 #include <list>
 #include <map>
@@ -127,8 +125,6 @@ std::string JSONRPCReply(const UniValue &result, const UniValue &error,
                          const UniValue &id);
 UniValue JSONRPCError(int code, const std::string &message);
 
-/** Get name of RPC authentication cookie file */
-fs::path GetAuthCookieFile();
 /** Generate a new RPC authentication cookie and write it to disk */
 bool GenerateAuthCookie(std::string *cookie_out);
 /** Read the RPC authentication cookie from disk */
