+++ src/validation.cpp
@@ -19,7 +19,6 @@
 #include "consensus/merkle.h"
 #include "consensus/tx_verify.h"
 #include "consensus/validation.h"
-#include "fs.h"
 #include "hash.h"
 #include "init.h"
 #include "policy/fees.h"
@@ -39,7 +38,7 @@
 #include "txmempool.h"
 #include "ui_interface.h"
 #include "undo.h"
-#include "util.h"
+#include "fs_util.h"
 #include "utilmoneystr.h"
 #include "utilstrencodings.h"
 #include "validationinterface.h"
