+++ src/wallet/walletdb.cpp
@@ -12,10 +12,10 @@
 #include "consensus/tx_verify.h"
 #include "consensus/validation.h"
 #include "dstencode.h"
-#include "fs.h"
 #include "protocol.h"
 #include "serialize.h"
 #include "sync.h"
+#include "fs_util.h"
 #include "util.h"
 #include "utiltime.h"
 #include "wallet/wallet.h"
