+++ src/wallet/walletutil.cpp
@@ -2,6 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "util.h"
 #include <wallet/walletutil.h>
 
 fs::path GetWalletDir() {
