+++ src/wallet/walletutil.h
@@ -2,13 +2,11 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#ifndef BITCOIN_WALLET_UTIL_H
-#define BITCOIN_WALLET_UTIL_H
+#pragma once
 
-#include <util.h>
+#include <fs_util.h>
 
 //! Get the path of the wallet directory.
 fs::path GetWalletDir();
 fs::path GetWalletDirNoCreate(fs::path& added_dir);
 
-#endif // BITCOIN_WALLET_UTIL_H
